backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
google-generativeai==0.8.0
firebase-admin==6.5.0
python-dotenv==1.0.1
qrcode==7.4.2
scipy==1.13.1
